For chunk_size >1GB calculate md5sums by portions.
Also moved the exceptions into their own module.
